refactor(src): further restructuring: - moved CLI specific code to a ... CLI folder ! (wow) - move convertToSolid to core - minor cleanups
