+++ src/cli/generateOutputData.js
